more bugs fixed related to jpa migration

+++ interfaces/src/main/java/ar/edu/itba/paw/services/AlbumService.java
@@ -18,6 +18,6 @@ public interface AlbumService extends CrudService<Album> {
     Album update(AlbumDTO albumDTO);
     boolean updateAll(List<AlbumDTO> albumsDTO, long artistId);
 
-    boolean updateRating(long albumId, float newRating, int newRatingAmount);
+    boolean updateRating(long albumId, Double newRating, int newRatingAmount);
     boolean hasUserReviewed(long userId, long albumId);
 }
